+++ media-api/app/MediaApiComponents.scala
@@ -26,13 +26,13 @@ class MediaApiComponents(context: Context) extends GridComponents(context) {
     replicas = config.elasticsearch6Replicas
   )
 
+  val elasticSearch = new lib.elasticsearch.impls.elasticsearch6.ElasticSearch(config, mediaApiMetrics, es6Config)
+  elasticSearch.ensureAliasAssigned()
+
   val s3Client = new S3Client(config)
 
-  val usageQuota = new UsageQuota(config, actorSystem.scheduler)
+  val usageQuota = new UsageQuota(config, elasticSearch, actorSystem.scheduler)
   usageQuota.scheduleUpdates()
-
-  val elasticSearch = new lib.elasticsearch.impls.elasticsearch6.ElasticSearch(config, mediaApiMetrics, es6Config, () => usageQuota.usageStore.overQuotaAgencies)
-  elasticSearch.ensureAliasAssigned()
 
   val imageResponse = new ImageResponse(config, s3Client, usageQuota)
