+++ media-api/app/lib/elasticsearch/impls/elasticsearch6/IsQueryFilter.scala
@@ -2,7 +2,6 @@ package lib.elasticsearch.impls.elasticsearch6
 
 import com.gu.mediaservice.lib.ImageFields
 import com.gu.mediaservice.model._
-import com.sksamuel.elastic4s.http.ElasticDsl.matchNoneQuery
 import com.sksamuel.elastic4s.searches.queries.Query
 import scalaz.syntax.std.list._
 
@@ -13,17 +12,15 @@ sealed trait IsQueryFilter extends Query with ImageFields {
     case IsOwnedPhotograph => "gnm-owned-photo"
     case IsOwnedIllustration => "gnm-owned-illustration"
     case IsOwnedImage => "gnm-owned"
-    case _: IsOverQuota => "over-quota"
   }
 }
 
 object IsQueryFilter {
   // for readability, the client capitalises gnm, so `toLowerCase` it before matching
-  def apply(value: String, overQuotaAgencies: () => List[Agency]): Option[IsQueryFilter] = value.toLowerCase match {
+  def apply(value: String): Option[IsQueryFilter] = value.toLowerCase match {
     case "gnm-owned-photo" => Some(IsOwnedPhotograph)
     case "gnm-owned-illustration" => Some(IsOwnedIllustration)
     case "gnm-owned" => Some(IsOwnedImage)
-    case "over-quota" => Some(IsOverQuota(overQuotaAgencies()))
     case _ => None
   }
 }
@@ -44,10 +41,4 @@ object IsOwnedImage extends IsQueryFilter {
   override def query: Query = filters.or(
     filters.terms(usageRightsField("category"), UsageRights.whollyOwned.toNel.get.map(_.category))
   )
-}
-
-case class IsOverQuota(overQuotaAgencies: List[Agency]) extends IsQueryFilter {
-  override def query: Query = overQuotaAgencies.toNel
-    .map(agency => filters.or(filters.terms(usageRightsField("supplier"), agency.map(_.supplier))))
-    .getOrElse(matchNoneQuery)
 }
